+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Each release is linked to a git commit.
 - Beta angle computation
 - Constellation and LEO builder
 - Local Time at Ascending Node computation module
+- Real adaptive step size for KeplerNum propagator
 
 ### Modified